feat: Move minimized calling view to the bottom of the conversation list
