+++ src/script/page/LeftSidebar/panels/Conversations/Conversations.tsx
@@ -51,7 +51,7 @@ import {EmptyConversationList} from './EmptyConversationList';
 import {getTabConversations} from './helpers';
 import {SidebarStatus, SidebarTabs, useFolderState, useSidebarStore} from './state';
 
-import {CallingViewMode, CallState} from '../../../../calling/CallState';
+import {CallState} from '../../../../calling/CallState';
 import {createLabel} from '../../../../conversation/ConversationLabelRepository';
 import {ConversationRepository} from '../../../../conversation/ConversationRepository';
 import {ConversationState} from '../../../../conversation/ConversationState';
@@ -120,9 +120,7 @@ const Conversations: React.FC<ConversationsProps> = ({
     'unreadConversations',
     'visibleConversations',
   ]);
-  const {activeCalls, viewMode} = useKoSubscribableChildren(callState, ['activeCalls', 'viewMode']);
-
-  const isCallWindowDetached = viewMode === CallingViewMode.DETACHED_WINDOW;
+  const {activeCalls} = useKoSubscribableChildren(callState, ['activeCalls']);
 
   const {conversationLabelRepository} = conversationRepository;
   const favoriteConversations = conversationLabelRepository.getFavorites(conversations);
@@ -278,20 +276,18 @@ const Conversations: React.FC<ConversationsProps> = ({
         const {callingRepository} = callingViewModel;
 
         return (
-          conversation &&
-          !isCallWindowDetached && (
-            <div className="calling-cell" key={conversation.id}>
-              <CallingCell
-                classifiedDomains={classifiedDomains}
-                call={call}
-                callActions={callingViewModel.callActions}
-                callingRepository={callingRepository}
-                pushToTalkKey={propertiesRepository.getPreference(PROPERTIES_TYPE.CALL.PUSH_TO_TALK_KEY)}
-                isFullUi
-                hasAccessToCamera={callingViewModel.hasAccessToCamera()}
-                isSelfVerified={selfUser.is_verified()}
-              />
-            </div>
+          conversation && (
+            <CallingCell
+              key={conversation.id}
+              classifiedDomains={classifiedDomains}
+              call={call}
+              callActions={callingViewModel.callActions}
+              callingRepository={callingRepository}
+              pushToTalkKey={propertiesRepository.getPreference(PROPERTIES_TYPE.CALL.PUSH_TO_TALK_KEY)}
+              isFullUi
+              hasAccessToCamera={callingViewModel.hasAccessToCamera()}
+              isSelfVerified={selfUser.is_verified()}
+            />
           )
         );
       })}
@@ -327,7 +323,7 @@ const Conversations: React.FC<ConversationsProps> = ({
         }
         hasHeader={!isPreferences}
         sidebar={sidebar}
-        before={callingView}
+        footer={callingView}
       >
         {isPreferences ? (
           <Preferences
